requirements.txt
Flask==3.0.0
google-generativeai==0.4.0
dlib==19.24.4
numpy==1.26.4
opencv-python==4.10.0.82
setuptools==70.0.0
python-dotenv
PyPDF2
flask-socketio
gtts